# Conflicts:
#	Test/GameObject.cpp
#	Test/GameOver.cpp
#	Test/main.cpp

+++ Test/MidPlatShieldPhysicsComponent.cpp
@@ -13,7 +13,7 @@ MidPlatShieldPhysicsComponent::MidPlatShieldPhysicsComponent(GameObject * platfo
 {
 	gameObjectRef = platform;
 	gameObjectRef->AddComponent(COMPONENT_PHYSICS, this);
-	init(size);
+	init((float)size);
 }
